Forks webserver by number of cpus
