+++ app/index.js
@@ -1,108 +1,42 @@
-console.log(require("helper").clfdate(), "app starting");
+const config = require("config");
+const cluster = require("cluster");
+const numCPUs = require("os").cpus().length;
+const clfdate = require("helper").clfdate;
 
-var config = require("config");
-var Express = require("express");
-var helmet = require("helmet");
-var vhost = require("vhost");
-var blog = require("./blog");
-var brochure = require("./brochure");
-var dashboard = require("./dashboard");
-var cdn = require("./cdn");
-var scheduler = require("./scheduler");
-var clfdate = require("helper").clfdate;
-var hash = require('helper').hash;
+if (cluster.isMaster) {
+  const scheduler = require("./scheduler");
 
-// Welcome to Blot. This is the Express application which listens on port 8080.
-// NGINX listens on port 80 in front of Express app and proxies requests to
-// port 8080. NGINX handles SSL termination, cached response delivery and
-// compression. See ../config/nginx for more. Blot does the rest.
-var Blot = Express();
+  console.log(clfdate(), `Master process ${process.pid} running`);
 
-// Removes a header otherwise added by Express. No wasted bytes
-Blot.disable("x-powered-by");
+  // Schedule backups, subscription renewal emails
+  // and the publication of scheduled blog posts.
+  scheduler();
 
-// Trusts secure requests terminated by NGINX, as far as I know
-Blot.set("trust proxy", "loopback");
-
-// Prevent <iframes> embedding pages served by Blot
-Blot.use(helmet.frameguard("allow-from", config.host));
-
-Blot.use(function(req, res, next) {
-  var init = Date.now();
-
-  try {
-    console.log(
-      clfdate(),
-      req.headers["x-request-id"],
-      req.method,
-      req.protocol + "://" + req.hostname + req.originalUrl,
-    );
-  } catch (e) {
-    console.error("Error: Failed to construct canonical log line:", e);
+  // Fork workers.
+  for (let i = 0; i < numCPUs; i++) {
+    cluster.fork();
   }
 
-  res.on("finish", function() {
-    try {
-      console.log(
-        clfdate(),
-        req.headers["x-request-id"],
-        res.statusCode,
-        ((Date.now() - init)/1000).toFixed(3),
-        req.protocol + "://" + req.hostname + req.originalUrl
-      );
-    } catch (e) {
-      console.error("Error: Failed to construct canonical log line:", e);
-    }
+  cluster.on("fork", (worker) => {
+    // console.log("worker is dead:", worker.isDead());
   });
 
-  next();
-});
+  cluster.on("exit", (worker, code, signal) => {
+    console.log("worker is dead:", worker.isDead());
+    cluster.fork();
+  });
 
-// Blot is composed of four sub applications.
+} else {
+  const Blot = require("./server");
 
-// The CDN
-// -------
-// Serves static files generated by Blot e.g.
-// thumbnails and cached images.
-Blot.use(vhost(config.host, cdn));
+  console.log(clfdate(), `Worker process ${process.pid} running`);
 
-// The Dashboard
-// -------------
-// Serve the dashboard and public site (the brochure)
-// Webhooks from Dropbox and Stripe, git pushes are
-// served by these two applications. The dashboard can
-// only ever be served for request to the host
-Blot.use(vhost(config.host, dashboard));
-
-// The Brochure
-// ------------
-// The least important application. It serves the documentation
-// and sign up page.
-Blot.use(vhost(config.host, brochure));
-
-// The Blogs
-// ---------
-// Serves the customers's blogs. It should come first because it's the
-// most important. We don't know the hosts for all the blogs in
-// advance so all requests hit this middleware.
-Blot.use(blog);
-
-// Monit, which we use to monitor the server's health, requests
-// localhost/health to see if it should attempt to restart Blot.
-// If you remove this, change monit.rc too.
-Blot.use("/health", function(req, res) {
-  res.send("OK");
-});
-
-// Open the server to handle requests
-Blot.listen(config.port, function() {
-  // The endpoints test listens for this line in stdout
-  // so if you change its text, ensure you update the
-  // START_MESSAGE variable in tests/app/endpoints.js
-  console.log(clfdate(), "app listening on port", config.port);
-  console.log(clfdate(), "app running in environment:", config.environment);
-});
-
-// Schedule backups, subscription renewal emails
-// and the publication of scheduled blog posts.
-scheduler();
+  // Open the server to handle requests
+  Blot.listen(config.port, function () {
+    // The endpoints test listens for this line in stdout
+    // so if you change its text, ensure you update the
+    // START_MESSAGE variable in tests/app/endpoints.js
+    console.log(clfdate(), "App listening on port", config.port);
+    console.log(clfdate(), "App running in environment:", config.environment);
+  });
+}
